implement max idle pings configuration
Defines a limit to the number of workers that can be pinged for each cycle of the `handle_ping/2` optional callback. Defaults to `nil`, which is no limit.

On large pools, if many resources goes idle at the same cycle you may end up terminating a large number of workers sequencially, what could lead to the pool being unable to fulfill requests.

The `:max_idle_pings` pool option is useful to prevent sequencial termination of a large number
  of workers. But it is important to keep in mind the following behaviours whenever utilizing it. See the added docs.